+++ HW-9/index.html
@@ -6,7 +6,9 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Document</title>
-  <link rel="stylesheet" href="./style.css" </head>
+  <link rel="stylesheet" href="./style.css">
+  <script src="./src/index.js" type="module"></script>
+</head>
 
 <body>
   <div class="container">
@@ -42,11 +44,13 @@
           <button type="button" class="button conroll-pannel_padding">X</button>
           <button type="button" class="button conroll-pannel_padding">Date</button>
         </div>
+      </div>
         -->
+
     </div>
   </div>
-  </div>
-  <script src="./main.js"></script>
+
+  <!-- <script src="./main.js"></script> -->
 </body>
 
 </html>
