+++ HW-9/src/index.js
@@ -1,0 +1,2 @@
+
+import "./app.js";
