+++ HW-9/src/store.js
@@ -1,0 +1,16 @@
+export const todoList = [
+  {
+    id: 0,
+    checked: true,
+    todoText: "first todo",
+    date: ``,
+  },
+  {
+    id: 1,
+    checked: false,
+    todoText: "second todo",
+    date: ``,
+  },
+]
+
+export const storageKey = "todos";
