+++ HW-9/src/utils/localStorage.js
@@ -1,0 +1,9 @@
+const setName = (storageKey, obj) => {
+  localStorage.setItem(storageKey, JSON.stringify(obj));
+}
+
+const getName = (storageKey) => {
+  return JSON.parse(localStorage.getItem(storageKey));
+}
+
+export { setName, getName };
